requirements.txt
Flask==2.1.3
pandas==1.4.3
waitress==2.1.2
python-dotenv==0.21.0
